% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_fdp}
\alias{compute_fdp}
\title{Compute False Discovery Proportion (FDP) using index sets}
\usage{
compute_fdp(selected_indices, true_support)
}
\arguments{
\item{selected_indices}{Integer vector of selected indices.}

\item{true_support}{Integer vector of true active indices.}
}
\value{
Numeric FDP value.
}
\description{
Compute False Discovery Proportion (FDP) using index sets
}
\examples{
compute_fdp(c(1L, 3L, 5L), c(1L, 2L, 3L))

}
