% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_precision}
\alias{compute_precision}
\title{Compute Precision using index sets}
\usage{
compute_precision(selected_indices, true_support)
}
\arguments{
\item{selected_indices}{Integer vector of selected indices.}

\item{true_support}{Integer vector of true active indices.}
}
\value{
Numeric precision value.
}
\description{
Compute Precision using index sets
}
\examples{
compute_precision(c(1L, 3L, 5L), c(1L, 2L, 3L))

}
