% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_recall}
\alias{compute_recall}
\title{Compute Recall using index sets}
\usage{
compute_recall(selected_indices, true_support)
}
\arguments{
\item{selected_indices}{Integer vector of selected indices.}

\item{true_support}{Integer vector of true active indices.}
}
\value{
Numeric recall value.
}
\description{
Compute Recall using index sets
}
\examples{
compute_recall(c(1L, 3L, 5L), c(1L, 2L, 3L))

}
