% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dim.mmap_matrix}
\alias{dim.mmap_matrix}
\title{Get mmap_matrix dimensions}
\usage{
\method{dim}{mmap_matrix}(x)
}
\arguments{
\item{x}{mmap_matrix object}
}
\value{
an integer vector of length 2
}
\description{
Get mmap_matrix dimensions
}
\examples{
\donttest{
mat <- matrix(as.double(1:20), nrow = 4, ncol = 5)
f <- tempfile(fileext = ".bin")
on.exit(unlink(f), add = TRUE)
convert_to_memory_mapped(mat, f)
mm <- mmap_matrix(f, rows = 4, cols = 5)
dim(mm)
}
}
