% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{LinkageMethod}
\alias{LinkageMethod}
\title{Linkage Methods}
\usage{
LinkageMethod
}
\description{
Enumeration of supported hierarchical clustering linkage methods.
Supported methods include:
- \code{Ward}: Ward's minimum variance method: codes as 0.
- \code{Average} (UPGMA): Average linkage: codes as 1.
- \code{Complete}: Complete linkage: codes as 2.
- \code{Single}: Single linkage: codes as 3.
- \code{WPGMA}: Weighted pair group method with arithmetic mean: codes as 4.
- \code{Median}: Median linkage: codes as 5.
- \code{Centroid}: Centroid linkage: codes as 6.
}
\examples{
LinkageMethod$Ward
LinkageMethod$Average
}
