% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsolvers.R
\name{TOOLSSolver}
\alias{TOOLSSolver}
\title{T-OLS Solver}
\description{
Terminating Orthogonal Least Squares (T-OLS) solver.
Inherits all common methods from \code{\link{TSolverBase}}.
}
\examples{
set.seed(1)
n <- 20; p <- 5
X <- matrix(rnorm(n * p), n, p)
D <- matrix(rnorm(n * p), n, p)
y <- X[, 1] * 2 + rnorm(n)
solver <- TOOLSSolver$new(X, D, y)
solver$execute_step(T_stop = 3)
solver$get_beta()

}
\section{Super class}{
\code{\link[TRexSelectorNeo:TSolverBase]{TSolverBase}} -> \code{TOOLSSolver}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-TOOLSSolver-initialize}{\code{TOOLSSolver$new()}}
    \item \href{#method-TOOLSSolver-clone}{\code{TOOLSSolver$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolverBase" data-id="execute_step"><a href='../../TRexSelectorNeo/html/TSolverBase.html#method-TSolverBase-execute_step'><code>TSolverBase$execute_step()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolverBase" data-id="get_actions"><a href='../../TRexSelectorNeo/html/TSolverBase.html#method-TSolverBase-get_actions'><code>TSolverBase$get_actions()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolverBase" data-id="get_actives"><a href='../../TRexSelectorNeo/html/TSolverBase.html#method-TSolverBase-get_actives'><code>TSolverBase$get_actives()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolverBase" data-id="get_beta"><a href='../../TRexSelectorNeo/html/TSolverBase.html#method-TSolverBase-get_beta'><code>TSolverBase$get_beta()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolverBase" data-id="get_cp"><a href='../../TRexSelectorNeo/html/TSolverBase.html#method-TSolverBase-get_cp'><code>TSolverBase$get_cp()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolverBase" data-id="get_dof"><a href='../../TRexSelectorNeo/html/TSolverBase.html#method-TSolverBase-get_dof'><code>TSolverBase$get_dof()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolverBase" data-id="get_inactives"><a href='../../TRexSelectorNeo/html/TSolverBase.html#method-TSolverBase-get_inactives'><code>TSolverBase$get_inactives()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolverBase" data-id="get_intercept"><a href='../../TRexSelectorNeo/html/TSolverBase.html#method-TSolverBase-get_intercept'><code>TSolverBase$get_intercept()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolverBase" data-id="get_num_steps"><a href='../../TRexSelectorNeo/html/TSolverBase.html#method-TSolverBase-get_num_steps'><code>TSolverBase$get_num_steps()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolverBase" data-id="get_r2"><a href='../../TRexSelectorNeo/html/TSolverBase.html#method-TSolverBase-get_r2'><code>TSolverBase$get_r2()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolverBase" data-id="get_rss"><a href='../../TRexSelectorNeo/html/TSolverBase.html#method-TSolverBase-get_rss'><code>TSolverBase$get_rss()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolverBase" data-id="load"><a href='../../TRexSelectorNeo/html/TSolverBase.html#method-TSolverBase-load'><code>TSolverBase$load()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolverBase" data-id="save"><a href='../../TRexSelectorNeo/html/TSolverBase.html#method-TSolverBase-save'><code>TSolverBase$save()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TOOLSSolver-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-TOOLSSolver-initialize}{}}}
\subsection{\code{TOOLSSolver$new()}}{
  Create a new \code{TOOLSSolver}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TOOLSSolver$new(X, D, y, normalize = TRUE, intercept = TRUE, verbose = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{X}}{Numeric matrix of predictors, or \code{mmap_matrix}.}
      \item{\code{D}}{Numeric matrix of dummies.}
      \item{\code{y}}{Numeric response vector.}
      \item{\code{normalize}}{Logical; column-normalize X and D (default \code{TRUE}).}
      \item{\code{intercept}}{Logical; fit an intercept (default \code{TRUE}).}
      \item{\code{verbose}}{Logical; print progress (default \code{FALSE}).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TOOLSSolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TOOLSSolver-clone}{}}}
\subsection{\code{TOOLSSolver$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TOOLSSolver$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
