% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_methods.R
\name{RidgeCV}
\alias{RidgeCV}
\title{Ridge Regression with Cross-Validation}
\description{
Performs explicit K-Fold cross-validation for Ridge Regression model selection.
}
\examples{
set.seed(1)
X <- matrix(rnorm(50 * 5), 50, 5)
y <- X[, 1] * 2 + rnorm(50)
cv <- RidgeCV$new()
cv$fit(X, y, num_folds = 5)
cv$cv_min()

}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-RidgeCV-initialize}{\code{RidgeCV$new()}}
    \item \href{#method-RidgeCV-fit}{\code{RidgeCV$fit()}}
    \item \href{#method-RidgeCV-cv_min}{\code{RidgeCV$cv_min()}}
    \item \href{#method-RidgeCV-cv_1se}{\code{RidgeCV$cv_1se()}}
    \item \href{#method-RidgeCV-index_min}{\code{RidgeCV$index_min()}}
    \item \href{#method-RidgeCV-index_1se}{\code{RidgeCV$index_1se()}}
    \item \href{#method-RidgeCV-get_lambdas}{\code{RidgeCV$get_lambdas()}}
    \item \href{#method-RidgeCV-get_cv_errors}{\code{RidgeCV$get_cv_errors()}}
    \item \href{#method-RidgeCV-get_cv_std}{\code{RidgeCV$get_cv_std()}}
    \item \href{#method-RidgeCV-clone}{\code{RidgeCV$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RidgeCV-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-RidgeCV-initialize}{}}}
\subsection{\code{RidgeCV$new()}}{
  Initialize the RidgeCV object.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RidgeCV$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RidgeCV-fit"></a>}}
\if{latex}{\out{\hypertarget{method-RidgeCV-fit}{}}}
\subsection{\code{RidgeCV$fit()}}{
  Fit the CV model.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RidgeCV$fit(
  X,
  y,
  num_folds = 10,
  n_lambda = 100,
  lambda_ratio = 1000,
  seed = 0
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{X}}{Predictor matrix.}
      \item{\code{y}}{Response vector.}
      \item{\code{num_folds}}{Number of folds (default 10).}
      \item{\code{n_lambda}}{Number of lambdas in path (default 100).}
      \item{\code{lambda_ratio}}{Ratio of max to min lambda (default 1000).}
      \item{\code{seed}}{Random seed for CV splits (default 0).}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Value.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RidgeCV-cv_min"></a>}}
\if{latex}{\out{\hypertarget{method-RidgeCV-cv_min}{}}}
\subsection{\code{RidgeCV$cv_min()}}{
  Get lambda that minimizes CV error.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RidgeCV$cv_min()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Value.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RidgeCV-cv_1se"></a>}}
\if{latex}{\out{\hypertarget{method-RidgeCV-cv_1se}{}}}
\subsection{\code{RidgeCV$cv_1se()}}{
  Get 1SE rule lambda.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RidgeCV$cv_1se()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Value.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RidgeCV-index_min"></a>}}
\if{latex}{\out{\hypertarget{method-RidgeCV-index_min}{}}}
\subsection{\code{RidgeCV$index_min()}}{
  Get index of minimum CV error.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RidgeCV$index_min()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Value.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RidgeCV-index_1se"></a>}}
\if{latex}{\out{\hypertarget{method-RidgeCV-index_1se}{}}}
\subsection{\code{RidgeCV$index_1se()}}{
  Get index of 1SE rule lambda.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RidgeCV$index_1se()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Value.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RidgeCV-get_lambdas"></a>}}
\if{latex}{\out{\hypertarget{method-RidgeCV-get_lambdas}{}}}
\subsection{\code{RidgeCV$get_lambdas()}}{
  Get the sequence of lambdas used.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RidgeCV$get_lambdas()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Value.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RidgeCV-get_cv_errors"></a>}}
\if{latex}{\out{\hypertarget{method-RidgeCV-get_cv_errors}{}}}
\subsection{\code{RidgeCV$get_cv_errors()}}{
  Get the CV mean squared errors for each lambda.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RidgeCV$get_cv_errors()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Value.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RidgeCV-get_cv_std"></a>}}
\if{latex}{\out{\hypertarget{method-RidgeCV-get_cv_std}{}}}
\subsection{\code{RidgeCV$get_cv_std()}}{
  Get the standard error of the CV errors.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RidgeCV$get_cv_std()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Value.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RidgeCV-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RidgeCV-clone}{}}}
\subsection{\code{RidgeCV$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{RidgeCV$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
