% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsolvers.R
\name{TLASSO_Solver}
\alias{TLASSO_Solver}
\title{T-LASSO Solver}
\description{
Terminating LASSO solver (T-LARS with variable removal).
Inherits all common methods from \code{\link{TSolver_Base}}.
}
\examples{
set.seed(1)
n <- 20; p <- 5
X <- matrix(rnorm(n * p), n, p)
D <- matrix(rnorm(n * p), n, p)
y <- X[, 1] * 2 + rnorm(n)
solver <- TLASSO_Solver$new(X, D, y)
solver$execute_step(T_stop = 3)
solver$get_beta()

}
\section{Super class}{
\code{\link[TRexSelectorNeo:TSolver_Base]{TSolver_Base}} -> \code{TLASSO_Solver}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-TLASSO_Solver-initialize}{\code{TLASSO_Solver$new()}}
    \item \href{#method-TLASSO_Solver-get_num_removals}{\code{TLASSO_Solver$get_num_removals()}}
    \item \href{#method-TLASSO_Solver-get_lambda}{\code{TLASSO_Solver$get_lambda()}}
    \item \href{#method-TLASSO_Solver-clone}{\code{TLASSO_Solver$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolver_Base" data-id="execute_step"><a href='../../TRexSelectorNeo/html/TSolver_Base.html#method-TSolver_Base-execute_step'><code>TSolver_Base$execute_step()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolver_Base" data-id="get_actions"><a href='../../TRexSelectorNeo/html/TSolver_Base.html#method-TSolver_Base-get_actions'><code>TSolver_Base$get_actions()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolver_Base" data-id="get_actives"><a href='../../TRexSelectorNeo/html/TSolver_Base.html#method-TSolver_Base-get_actives'><code>TSolver_Base$get_actives()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolver_Base" data-id="get_beta"><a href='../../TRexSelectorNeo/html/TSolver_Base.html#method-TSolver_Base-get_beta'><code>TSolver_Base$get_beta()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolver_Base" data-id="get_cp"><a href='../../TRexSelectorNeo/html/TSolver_Base.html#method-TSolver_Base-get_cp'><code>TSolver_Base$get_cp()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolver_Base" data-id="get_dof"><a href='../../TRexSelectorNeo/html/TSolver_Base.html#method-TSolver_Base-get_dof'><code>TSolver_Base$get_dof()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolver_Base" data-id="get_inactives"><a href='../../TRexSelectorNeo/html/TSolver_Base.html#method-TSolver_Base-get_inactives'><code>TSolver_Base$get_inactives()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolver_Base" data-id="get_intercept"><a href='../../TRexSelectorNeo/html/TSolver_Base.html#method-TSolver_Base-get_intercept'><code>TSolver_Base$get_intercept()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolver_Base" data-id="get_num_steps"><a href='../../TRexSelectorNeo/html/TSolver_Base.html#method-TSolver_Base-get_num_steps'><code>TSolver_Base$get_num_steps()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolver_Base" data-id="get_r2"><a href='../../TRexSelectorNeo/html/TSolver_Base.html#method-TSolver_Base-get_r2'><code>TSolver_Base$get_r2()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolver_Base" data-id="get_rss"><a href='../../TRexSelectorNeo/html/TSolver_Base.html#method-TSolver_Base-get_rss'><code>TSolver_Base$get_rss()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolver_Base" data-id="load"><a href='../../TRexSelectorNeo/html/TSolver_Base.html#method-TSolver_Base-load'><code>TSolver_Base$load()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="TRexSelectorNeo" data-topic="TSolver_Base" data-id="save"><a href='../../TRexSelectorNeo/html/TSolver_Base.html#method-TSolver_Base-save'><code>TSolver_Base$save()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TLASSO_Solver-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-TLASSO_Solver-initialize}{}}}
\subsection{\code{TLASSO_Solver$new()}}{
  Create a new \code{TLASSO_Solver}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TLASSO_Solver$new(X, D, y, normalize = TRUE, intercept = TRUE, verbose = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{X}}{Numeric matrix of predictors, or \code{mmap_matrix}.}
      \item{\code{D}}{Numeric matrix of dummies.}
      \item{\code{y}}{Numeric response vector.}
      \item{\code{normalize}}{Logical; column-normalize X and D (default \code{TRUE}).}
      \item{\code{intercept}}{Logical; fit an intercept (default \code{TRUE}).}
      \item{\code{verbose}}{Logical; print progress (default \code{FALSE}).}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TLASSO_Solver-get_num_removals"></a>}}
\if{latex}{\out{\hypertarget{method-TLASSO_Solver-get_num_removals}{}}}
\subsection{\code{TLASSO_Solver$get_num_removals()}}{
  Get the number of coefficient removals.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TLASSO_Solver$get_num_removals()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Integer count of removals.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TLASSO_Solver-get_lambda"></a>}}
\if{latex}{\out{\hypertarget{method-TLASSO_Solver-get_lambda}{}}}
\subsection{\code{TLASSO_Solver$get_lambda()}}{
  Get the regularization (lambda) path.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TLASSO_Solver$get_lambda()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Numeric vector of lambda values, one per step.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TLASSO_Solver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TLASSO_Solver-clone}{}}}
\subsection{\code{TLASSO_Solver$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TLASSO_Solver$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
